# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        description: уникальный индекс
        type: string
      id:
        description: UUID primary key
        type: string
      isVerified:
        type: boolean
      passwordHash:
        type: string
      resetToken:
        type: string
      resetTokenExp:
        type: string
      role:
        description: model | employer | admin
        type: string
      status:
        description: pending | active | suspended | banned
        type: string
      subscription:
        description: free | premium | pro
        type: string
      updatedAt:
        type: string
      verificationToken:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Модельное агенство
    url: http://contact.company.com
  description: API для управления модельной компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: mwork API
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: failed to delete user
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Возвращает пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: user not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить пользователя по ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет поля пользователя по его ID (например, статус)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: update
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: failed to update user
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновить пользователя
      tags:
      - users
swagger: "2.0"